+++ protected/humhub/modules/user/messages/de/views_auth_createAccount.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   '<strong>Account</strong> registration' => '<strong>Konto</strong>-Registrierung',
-  'Create Account' => 'ein Konto anlegen',
+  'Create Account' => 'Ein Konto anlegen',
 );