+++ protected/humhub/modules/user/tests/codeception/fixtures/data/group_user.php
@@ -20,5 +20,6 @@
 return [
     ['id' => 1, 'user_id' => 1, 'group_id' => 1, 'created_at' => '2014-08-30 14:03:49', 'created_by' => NULL, 'updated_at' => NULL, 'updated_by' => NULL],
     ['id' => 2, 'user_id' => 2, 'group_id' => 2, 'created_at' => '2014-08-30 14:03:49', 'created_by' => NULL, 'updated_at' => NULL, 'updated_by' => NULL],
-    ['id' => 3, 'user_id' => 3, 'group_id' => 3, 'created_at' => '2014-08-30 14:03:49', 'created_by' => NULL, 'updated_at' => NULL, 'updated_by' => NULL]
+    ['id' => 3, 'user_id' => 3, 'group_id' => 3, 'is_group_manager' => '1', 'created_at' => '2014-08-30 14:03:49', 'created_by' => NULL, 'updated_at' => NULL, 'updated_by' => NULL],
+    ['id' => 4, 'user_id' => 6, 'group_id' => 3, 'created_at' => '2014-08-30 14:03:49', 'created_by' => NULL, 'updated_at' => NULL, 'updated_by' => NULL]
 ];